feat: Move Backup & Cache Repository to new folder structure
