+++ src/script/main/app.ts
@@ -34,6 +34,9 @@ import {WebAppEvents} from '@wireapp/webapp-events';
 import {PrimaryModal} from 'Components/Modals/PrimaryModal';
 import {AssetRepository} from 'Repositories/assets/AssetRepository';
 import {AudioRepository} from 'Repositories/audio/AudioRepository';
+import {BackupRepository} from 'Repositories/backup/BackupRepository';
+import {BackupService} from 'Repositories/backup/BackupService';
+import {CacheRepository} from 'Repositories/cache/CacheRepository';
 import {initializeDataDog} from 'Util/DataDog';
 import {DebugUtil} from 'Util/DebugUtil';
 import {Environment} from 'Util/Environment';
@@ -47,9 +50,6 @@ import {AppInitializationStep, checkIndexedDb, InitializationEventLogger} from '
 import '../../style/default.less';
 import {SIGN_OUT_REASON} from '../auth/SignOutReason';
 import {URLParameter} from '../auth/URLParameter';
-import {BackupRepository} from '../backup/BackupRepository';
-import {BackupService} from '../backup/BackupService';
-import {CacheRepository} from '../cache/CacheRepository';
 import {CallingRepository} from '../calling/CallingRepository';
 import {ClientRepository, ClientService} from '../client';
 import {getClientMLSConfig} from '../client/clientMLSConfig';
